Split up honkbot class into separate classes with their own responsibilities
This is just to make it easier to parse for future people when looking
at honkbot. This is just breaking up the single monolithic class into
individual classes, each with their own responsibilities.

The advantage here is that you don't have to remember _where_ in the
code you need to modify something. You can just go right to the file and
do the thing.

It's not as clean as I'd like it, but it's a good first pass.